bump beancount-mode
moving to the major mode required a bit of fiddling to get comfortable.
i had to upgrade swiper at the same time because i was on an ancient
version of counsel, and counsel-outline didn't know how to display full
paths to headings at that time. surprisingly, none of my other
swiper/ivy/counsel workflows broke