improve counsel-projectile configuration
huzzah, there's a function that does the awful index manipulation i used
to do to the action list! also, counsel-projectile-find-file will prompt
you for a project if you aren't in one now. i'm pretty sure it used to
just error out before, which is why i needed the custom function. not
anymore!