update projectile ignore/unignore handling
projectile 2.0.0 changed the default indexing method to ignore all
elisp-induced modifications. it uses whatever the alien program
returned, verbatim. this means all the ignore/unignore/.projectile
adjustments are irrelevant. on the whole i think this is a good thing,
but i want to be able to override it gracefully in my dir locals (eg for
projects that contain a lot of binary junk that i don't want to see in
my find-file list).