minimum updates for emacs29
use-package is integrated now. i also tried a fresh clone and had to
bump melpa because the recipe for emacs-undo-fu was using a gitlab repo
that cannot be cloned anonymously (the current melpa recipe for that
package clones it from codeberg instead). likely more tweaks and package
updates to follow this.