disable use-package-use-theme
when i removed use-package from my straight packages, it upgraded to the
vendored version in emacs, which brought in quite a few changes,
including this custom theme thing. for some reason, two of evil's
defcustoms do not get set when this theme stuff is used. so i just turn
it off